+++ R/visNetwork.R
@@ -3,14 +3,20 @@
 # Note: legend is not used right now!
 .plot.visNetwork <-
   function(x,
-    belief = TRUE,
+    belief = NULL,
+    show_belief = TRUE,
     legend = NULL,
     col = NULL,
     smooth = list(type = "continuous"),
+    layout = NULL,
     ...) {
     check_installed("visNetwork")
-    
-    pg <- policy_graph(x, belief = belief, col = col)
+   
+    unconverged <-  !x$solution$converged || length(x$solution$pg) > 1
+    if (is.null(layout))
+      layout <- ifelse(unconverged, "layout_as_tree", "layout_nicely")
+     
+    pg <- policy_graph(x, belief, show_belief = show_belief, col = col)
     
     ### add tooltip
     #V(pg)$title <- paste(htmltools::tags$b(V(pg)$label)
@@ -23,11 +29,11 @@
       ))
     
     ### colors
-    if (belief) {
+    if (show_belief) {
       # winner
       #V(pg)$color <- V(pg)$pie.color[[1]][sapply(V(pg)$pie, which.max)]
       
-      # mixing in rgb spave
+      # mixing in rgb space
       V(pg)$color <- sapply(
         seq(length(V(pg))),
         FUN = function(i)
@@ -41,7 +47,7 @@
       #  do.call(hsv, as.list(rgb2hsv(col2rgb(V(pg)$pie.color[[1]])) %*% V(pg)$pie[[i]])))
     }
     
-    visNetwork::visIgraph(pg, idToLabel = FALSE, smooth = smooth, ...) %>%
+    visNetwork::visIgraph(pg, idToLabel = FALSE, layout = layout, smooth = smooth, ...) %>%
       visNetwork::visOptions(
         highlightNearest = list(enabled = TRUE, degree = 0),
         nodesIdSelection = TRUE